arm: Use a consistent type (uintptr_t) for g_idle_topstack

+++ arch/arm/src/cxd56xx/cxd56_allocateheap.c
@@ -69,7 +69,7 @@
  * aligned).
  */
 
-const uint32_t g_idle_topstack = (uint32_t)&_ebss +
+const uintptr_t g_idle_topstack = (uintptr_t)&_ebss +
   CONFIG_IDLETHREAD_STACKSIZE;
 
 /****************************************************************************
